Refactor infra files
