+++ src/CLIPreTasks/CLIPreTasks.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk'
 import { pathExistsSync, removeSync } from 'fs-extra'
 import { join } from 'path'
 import semver from 'semver'
-import { configDir } from '../conf'
+import { configDir } from '../api/conf'
 import { PathConstants } from '../lib/constants/Paths'
 import { DeprecationChecker } from './DeprecationChecker/DeprecationChecker'
 import { OutdatedChecker } from './OutdatedChecker/OutdatedChecker'
