+++ src/api/index.ts
@@ -12,5 +12,6 @@ export * from './modules'
 export * from './table'
 export * from './verbose'
 export * from './constants'
+export * from './conf'
 export { BatchStream } from './typings'
 export { default as logger } from './logger'
