+++ src/oclif/hooks/init.ts
@@ -11,7 +11,7 @@ import { hrTimeToMs } from '../../lib/utils/hrTimeToMs'
 import { updateNotify } from '../../update'
 import log from '../../api/logger'
 import * as pkg from '../../../package.json'
-import * as conf from '../../conf'
+import * as conf from '../../api/conf'
 import { checkAndOpenNPSLink } from '../../nps'
 import { Metric } from '../../api/metrics/MetricReport'
 import authLogin from '../../modules/auth/login'
